Finish ghost movment
The ghost does not use a path to navigate. The ghost decide randomly after every direction changing or when the ghost can not go in the new calculated direction.
Moreover the ghosts are managed in the GameState class in an array, so adding more ghosts to the game is easier.
The spawing position and the number of ghosts are now also defined in the World1.txt